src/requirements.txt
pymysql==0.10.0
python-dotenv==0.14.0